MercadoPagoUIScrollViewController introduced. RyC main title added as section.
